refactor : 이미지 업로드 api 분리
