+++ frontend/src/entities/textEdit/index.ts
@@ -1,2 +1,3 @@
 export { headerInput } from './ui/headerInput';
 export { bottomButtons } from './ui/bottomButtons';
+export { uploadImage } from './api/uploadImage';
