+++ frontend/src/entities/textEdit/ui/ImageUpload.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 import imgfile from '/src/shared/assets/image/markdownImg/Image.svg';
-import axios from 'axios';
+import * as textEdit from '../index';
 
 export const ImageUpload = ({ setImageurl }: any) => {
   const [selectedImage, setSelectedImage] = useState<File | null>(null);
@@ -19,29 +19,7 @@ export const ImageUpload = ({ setImageurl }: any) => {
   };
 
   useEffect(() => {
-    const uploadImage = async () => {
-      if (selectedImage) {
-        try {
-          formData.append('multipartFile', selectedImage);
-
-          const response = await axios.post('/api/image/upload', formData, {
-            headers: {
-              'Content-Type': 'multipart/form-data',
-              authorization:
-                'Bearer eyJhbGciOiJIUzUxMiJ9.eyJpYXQiOjE3MTQ2NTY0OTUsImV4cCI6MTcxNTg2NjQ5NX0.jiNtmKxvv75AMh0IaikVR-6kkJ4UZbrvHEi_oxgphLCv7wUWH75QDN3gi0UwcUYecPhQfH5XbZknXU6aOVdopw',
-            },
-          });
-
-          setSelectedImage(null);
-          setImageurl(`![](${response.data.data})`);
-        } catch (error) {
-          console.log(error);
-          alert('더 작은 용량의 이미지를 업로드해주세요!');
-        }
-      }
-    };
-
-    uploadImage();
+    textEdit.uploadImage(selectedImage, formData, setSelectedImage, setImageurl);
   }, [selectedImage]);
   return (
     <>
